Add Functionality of the following--
Email Verification in app as well as via API,that email is allready
registered in SSID or not.
Phone No Verification in app as well as via API and Firebase otp,that
Phone no is allready registered in SSID or not and No is valid or not.
Password validation is implemented.
After Authentication passing the value to Run2PLay.